🔨 Rename simulation trigger
